UpdatePageCoreのリファクタに付随する修正
